+++ includes/blocks.php
@@ -26,7 +26,10 @@ function register_dynamic_blocks() {
 		return;
 	}
 
-	$block_names = array( 'list-courses.php' => 'hrscourses/list-courses' );
+	$block_names = array(
+		'list-courses.php' => 'hrscourses/list-courses',
+		'courses-list.php' => 'hrscourses/courses-list',
+	);
 
 	foreach ( $block_names as $file => $block_name ) {
 		if ( ! file_exists( $blocks_dir . $file ) ) {